requirements.txt
python-dotenv
fastapi
sqlalchemy
pymysql
pydantic
argon2-cffi
python-jose[cryptography]
uvicorn